+++ Encoding/QSVEnc.vb
@@ -406,7 +406,11 @@ Public Class QSVEnc
                     sourcePath = "-"
 
                     If includePaths Then
-                        ret = If(includePaths, Package.ffmpeg.Path.Escape, "ffmpeg") + " -threads 1 -hwaccel dxva2 -i " + If(includePaths, p.LastOriginalSourceFile.Escape, "path") + " -f yuv4mpegpipe -strict -1 -pix_fmt yuv420p -loglevel fatal -hide_banner - | " + If(includePaths, Package.QSVEnc.Path.Escape, "QSVEncC64")
+                        Dim pix_fmt = If(p.BitDepth = 10, "yuv420p10le", "yuv420p")
+                        ret = If(includePaths, Package.ffmpeg.Path.Escape, "ffmpeg") + " -threads 1 -hwaccel dxva2 -i " +
+                            If(includePaths, p.LastOriginalSourceFile.Escape, "path") + " -f yuv4mpegpipe -pix_fmt " +
+                            pix_fmt + " -strict -1 -loglevel fatal -hide_banner - | " +
+                            If(includePaths, Package.QSVEnc.Path.Escape, "QSVEncC64")
                     End If
                 Case "ffqsv"
                     sourcePath = "-"